Drop `after_success` from the build section.
Previously if an upload failed on Travis CI, the build still passed.
This is because `after_success` ignores any exit code (even non-zero
ones). As a result, it wouldn't be obvious that an upload failed or that
one should check at all. Moving the upload section into the `script`
section should at least ensure that if the upload fails the exit status
will be respected and thus fail the build. As a result, users should be
able to see this failure clearly.