circle tmpl: Checkout the merge commit of the PR if it is a PR.
Unlike other CIs, CircleCI does not test the merge commit of a PR with
its base branch. Instead it tests the PR's head commit. The problem with
this is the PR's status could differ when it is merged with the base
branch. For instance, if one needs changes in the base branch to get
their PR to pass, they must rebase/merge to get them into the history
the PR. This normally isn't a problem, but sometimes things do go wrong
when doing this merge/rebase, which adds a new unneeded difficulty.
Alternatively, a passing PR could turn out to fail when merged into the
base branch because some new content in the base branch was not tested
against in the PR.

To solve these issues, we checkout the merge ref for a PR (if it is a
PR) from GitHub. However, it should be noted that the merge ref can be
out-of-date in some cases w.r.t. the base branch. Still this is the
commonly used strategy on Travis CI and AppVeyor. If we had enough info,
we could ideally terminate a build that has merge conflicts.
Unfortunately it doesn't seem that CircleCI gives us this info. However,
now that the linter complains about merge conflicts. This should provide
end users the needed info to determine a CircleCI build in those cases
is unreliable.