Remove some now unneeded IO utility functions.
The functions dropped include the following.

* `get_file_blob`
* `get_mode_file`
* `set_mode_file`

It was found that the manipulations to git blobs to change permissions
had some unintended side effects. Namely it could effect permissions on
other unrelated files. As a result, `set_exe_file` was written to use
the `git` subcommand `update-index` to change these permissions on
specific files instead.