Explicitly remove the `defaults` channel.
It is possible that a user may not want the `defaults` channel. This may
be desired as they wish to build a clean stack from scratch. Though our
main concern is that including `defaults` implicitly or hard-coding its
inclusion can result in differences between re-rendering and build time
that will cause a variety of subtle bugs. Given `defaults` is included
if the source channels are not specified, there is no harm in us
dropping it in this way.

Here we drop the hard-coding of `defaults` from the docker build
`.condarc`. Also we explicitly remove `defaults` from the `.condarc` on
Travis CI and AppVeyor before proceeding to add channels. As a result,
we can be assured that only the channels that we specify in
`conda-forge.yml` or the default set included in `conda-smithy` will be
the only channels used as sources and in the order we have specified.