Add support for Python 3.6.
While technically re-rendering did work for Python 3.6, the Miniconda
selected for Python 3.6 on AppVeyor was not the ideal choice. Namely it
used the Python 2.7 one, which is built on VC9. Instead we change this
so that Python 3.6 leverages the Python 3.5 Miniconda install, which
uses the same VC14 runtime. Having the `root` Python version using the
same VC as the one used in a build has tended to work well for us in the
past and has generally been the strategy used. This returns us to that
strategy for Python 3.6.