Clean up various AppVeyor hacks and kludge.
Some of the changes included in here install a very particular version
of `conda`. Since then newer version have been released that we use
successfully. So it makes sense to remove the pinning. However, the
update is still needed as AppVeyor's included `conda` is so old.

Also we found ourselves in a situation where Python 2.7 jobs where
running out of memory with particular configurations of the root
environment when running `conda-forge-build-setup`. Some hacks were
added to adjust the `root` environment to fix this. Though these appear
to be no longer necessary.