Use `defaults` for the tests
If the channels are not specified, `get_index` will fill them in from
info it finds in `.condarc`. To avoid getting a potentially incorrect
result due to local settings, simply specify that we only want to check
`defaults`. We know there are many differences between `osx-64` and
`win-64` on `defaults` including `vc` features in package names, MSYS2
packages (from `m2`), etc. So we can be confident that if these match up
something went terribly wrong.