Travis: Split conda install and configure steps
Install Miniconda in the first step and then update and configure the
`root` environment in the second step. This will make way for other
optimizations later on even though this is not an optimization itself.
Also explain a bit more about what is going on with Travis CI builds,
add some nice messages in the logs to clarify what these steps are
doing.