Explicitly disable secrets in forked PRs
Seems that CircleCI was using a deploy key with forked PR builds.
However, they were not exposing the secret environment variables before.
Though now it appears that they are exposing secret environment
variables too. This is a bit strange as the CircleCI defaults to having
the setting for secret environment variables disabled by default. In
other words, it appears this default is not respected by CircleCI
currently.

As a workaround, we seem to be able to fix this by explicitly disabling
secret environment variables as well. A little bit of testing verifies
this works as intended. While it is a workaround, it is a good thing to
keep should CircleCI ever change the default value of this setting. Thus
we have added this change to the configuration of CircleCI on new
feedstocks right after project creation in this commit.

Should note this also uses a non-public portion of the API as does
enabling PRs from forks. However, we have gotten assurances from
CircleCI before that this is ok to use.