Combine CircleCI re-rendering steps
Previously the re-rendering of `ci_support/run_docker_build.sh` and
`circle.yml` occurred in two different steps. This combines these two
steps into one. Meaning one less retrieval from the index to determine
the contents of the matrix. As a result, there are exactly as many calls
to the index as there are target architectures. This should speed up
re-rendering a little bit and simplify the codebase a bit as well.