Consolidate Jinja branches in CircleCI config file
As the filtered branches having been moved closer to the build step in
the CircleCI config file, having two Jinja `if`-clauses no longer makes
sense. So this groups them together with the CircleCI skip fallback
test. Also this is done such that the CircleCI config file should not
change due to this difference in the template.