Configure CircleCI config file for matrix support
Update the CircleCI configuration file, `.circleci/config.yml`, to
specify the different jobs in the matrix and set the relevant
environment variables for each build (if any). This results in no change
for the skipped Linux build case. Also if there are no environment
variables, that step is never added to the configuration file.