Add a hook for feedstocks updates
Provides a hook to listen for push events on feedstocks. The intent
being to capture these events and use them to update each feedstock in
the `feedstocks` repo. The net result being that the `feedstocks` repo
will stay more up-to-date with feedstock additions in the organization
as well changes made to each feedstock. Since these events will only
concern an individual feedstock at a time, this should cutdown on the
number of calls made to the GitHub API for operations like this one.