Extend linter to check shell scripts via shellcheck
As the shell script syntax dates back to the 1970s, there are plenty of
pitfalls and old types of syntax that won't get removed due to backward
compatibility, which shellcheck can help to find for us.

Especially the build.sh is an important companion to the meta.yaml to
build complex recipes.