Add clone_depth parameter that sets git depth for all providers.
By default, clone_depth is None which results in no change to the CI
provider defaults (and no change to the rendered CI files). If
clone_depth is set to 0, a complete git clone with no depth limit will
be done (this was already the default in some cases, but not all).
A non-zero clone_depth will clone to that depth for all providers.