Fix upload for Windows packages using github_actions provider.
With `set "GIT_BRANCH=%GITHUB_REF:*/=%"`, GIT_BRANCH was coming out as
heads/<branchname>. The "refs/heads/" prefix is a constant according to
https://docs.github.com/en/actions/reference/context-and-expression-syntax-for-github-actions#github-context
It seems safe to just strip that instead.

The `if %UPLOAD_PACKAGES% == "True"` statement was failing for two
reasons: 1) Case sensitivity with "%UPLOAD_PACKAGES%" evaluating to
"true", and 2) Use of quotation marks needing to match on both sides of
the equals (either included or not). This fixes the case and makes the
match insensitive to boot, and adds quotation marks to the left side.